+++ src/Config.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace PhpStyler;
 
-use PhpStyler\Format\Format;
+use PhpStyler\Format\AFormat;
 use PhpStyler\Format\PlainFormat;
 
 class Config
@@ -14,7 +14,7 @@ class Config
     public function __construct(
         public readonly iterable $files,
         public readonly ?string $cache,
-        public readonly Format $format = new PlainFormat(),
+        public readonly AFormat $format = new PlainFormat(),
     ) {
     }
 }